Updating the format of rawstatistics local config
The new `Configuration` class requires that the type of merged values to be the
same. It is also desirable that the order in which entries in the `realms`
property retain the order in which they are read / processed. To this end the
`realms` value has been changed to an array in the base config file. This commit
propagates that change to its local config file.