Update application categorization for pegasus.
The pegasus software was incorrectly marked as proprietary. This change updates
the application identification info so that new ingests will have the correct
identification. The migration file updates the existing records to change the
ident to the correct one.

The next time aggregation runs it will automatically re-aggregate the rows that
need to be redone.