Add drilldown to jobs for the JobEfficiency component.
Clicking on a person/pi in the job efficiency component
now brings up a window that lists the jobs that were classified
as inefficient for the person/pi.

While implementing this feature two seperate bugs were found:
1) The aggregation query erronously omitted the job_category_id from the group by
   leading to incorrect values in the job_category_id column
2) The new json config in the group-by refactor had the wrong table name for the
   category groupby

these two bugs are fixed in this pull request because you need the changes
in this pull request to be able to confirm that the fixes work.
fixed at the same time since there was previously noway to test the bug