Make display of "Exit Status" filters deterministic.
The original code used the id value to sort the exit status filters.
Since the exit status is added to dynamically as value are ingested this
means that the sort order can vary between xdmod instances.

This changes it to use the name of the exit status so that the filter
order is determined bvy the sql collation.