Add ability to disable the analytze table step from the aggregate_supremm.sh script
The job performance data aggregation step runs the mysql analyze tables
command by default after aggregation. This command can take a long time
to run and the XDMoD user interface is unavailable while the commmand is
running. This change updates the top level aggregate_supremm.sh to add
a command line flag to disable the analyze table step.

It is recommended that the analyze tables command should be run periodically
but does not need to be run every day. At CCR the anaylze tables step is run
once per week on a weekend day.