Checking compatability w/ the new xdmod-qa changes
- Moving travis post-install-tests and updating shippable to match

- Since we're no longer using Travis it didn't make sense to keep the
`post-install-test.sh` script in a `.travis` folder. I've moved this script to
the same location that we keep our other ci scripts.

- Also updated the `post-install-test.sh` script so that it takes into account
that it's now operating on an rpm install.

- The running of the `post-install-test.sh` script has been added to shippable.yml
w/ the addition of the reseting of the `XDMOD_INSTALL_DIR` env variable to
reflect where XDMOD is actually installed. This was needed because the travis
tests test that it can be built & installed to a directory but this is done via
the "source" install and not the rpm install that we're using in shippable. So,
we have it build / install to `/xdmod` while the rpm is installed to
`/usr/share/xdmod`.