Add a walltime accuracy analytic and update documentation.
The walltime accuracy analytic is added bringing the total number up
to five. There is a corresponding change to the core to allow the
job viewer to display these: https://github.com/ubccr/xdmod/pull/1517

A new documentation page is also added to explain how to configure
which analytics are displayed. This is obv. not the full configurabilty
that we ultimately would like to have, but at least should head-off
some of the more common support requests.