Handle case where ingest/aggregation taskes a long time.
If the ingest or aggregation step takes a long time then the database
connection in the Journal Helper times out and the journal entry
does not get written.

This change explicity closes the database connection after each read/write.
The code will automatically open a new connection when needed. The other
possible change would have been to catch the server-has-gone-away exception and
retry (which is what I've done for the logger).  In this case the usage pattern
is well known: the helper needs to access the database once at the start and
once at the end and not in the middle so there is no need to keep an idle
connection open.