Use the git-provided commit range.
The QA scripts will use the environment variable TRAVIS_COMMIT_RANGE to
determine which commits to use when running the linter on the changed code. If
this environment variable is not set then the script will attempt to diff
against upstream/HEAD.

The upstream/HEAD diff is a problem when trying to ci test against old branches
since the diff will incorrectly compare the set of changes against the latest
build and not the correct target branch.

This changes the code to set the commit range appropriately so the scripts diff
against the correct commit.

See also the corresponding change in the main repo: https://github.com/ubccr/xdmod/pull/1552