Updating Resource GroupBys Underlying SQL
There was a visual bug present when adding a Resource filter for SUPREMM data
where multiple instances of the same resource name would be present in the chart
subtitle. This was caused by the underlying SQL joining `resourcefact` to
`resourcespecs` which is a one (resourcefact) -> many (resourcespecs)
relationship. This is due to resourcespecs being a historical table that
captures the change over time that a resource goes through. The most current
spec for a resource is identified by the one that has a null `end_date_ts`
value.