Add the end_time of the job to the lookup for tg_job_id
This is a redo of #284 on the xdmod 10.0 branch.

There is a missing where condition on the lookup that matches a job in the supremm realm to the jobs realm. In the event that job identifiers get reused by the resource manager and there is more than one job with the same job identifier then the information about the job in the supremm realm may be incorrectly mapped to the wrong job (with the same identifier) but that ran at a different time.