Update version of nodejs to the current LTS.
This changes the RPM dependencies to require nodejs 16 and removes
npm as a dependency (npm is packaged up in the nodesource rpm for
nodejs and also I'm not sure that we even need it!).

Added an extra script for the CI build to make sure the docker
is updated with the correct dependency.