Increase composer file cache timeout.
The default timout for files in the composer file cache is 180 days.
Its been a while since we updated some of these dependencies so
composer has been ageing them out and re-downloading. We noticed
that this was happening because there was a seperate issue with
downloading the highcharts dependency files.

This change increases the composer cache timeout so that the cached
files are perferred. This will (slightly) improve the build times
and works-around the (hopefully temporary) problem with the highcharts
download.