Update rpm deps and docs for nodejs EPEL version.
Revert CI build to use the EPEL version of nodejs. Note the previous
RPM had a runtime dependency on npm. This is not actually needed it
is really a build time dependency. The spec file has been updated to
fix this.

Also update the version number in the markdown (should have been done
earlier).