Removing unneeded step in circleci config
The conditions that required this step are now different and so it can
be removed.

The conditions were that originally xdmod10.0's composer.json file
didn't have the `composer/installers` and
`simplesamlphp/composer-module-installer` in the `allow-plugins` config
so the build would hang as it would interactively prompt to allow these
plugins.