Adding some regex slashify-ing per @jtpalmer
@jtpalmer caught that the regex being passed to the old MongoRegex class
was not valid, as it had been updated to what is required for the new
Mongo\BSON\Regex class, and throwing exceptions which was causing the
integration tests to fail.

This commit adds a helper function that will
add a trailing / leading slash if one is missing from the regex to make
it compliant with what \MongoRegex expects.